Fix a signedness bug when shifting bits; this caused the length field of the packet to occasionally been misinterpreted as a negative number.
When discarding a packet for being overly long, restart the state
machine instead of attempting to drop a preposterous amount of data.
It is unlikely in that case that preposterous amount of data would
ever arrive, so rather attempt to re-align the reading algorithm
(supposedly resulting in a timeout and retransmit).